Directly use <Motion/> if not reducing motion
